requirements/tests.txt
flake8
django_nose
coverage
selenium
